Namespace Adjustments

+++ Germac.Domain/Repositories/IRepository.cs
@@ -2,10 +2,10 @@
 {
     public interface IRepository<T>
     {
-        Task<T> GetById(string query, int id);
+        Task<T> GetById(string query, long id);
         Task<IEnumerable<T>> GetAll(string query);
         Task<int> Add(string query, T entity);
         Task<int> Update(string query, T entity);
-        Task<int> Delete(string query, int id);
+        Task<int> Delete(string query, long id);
     }
 }
